chore: rename binary package to `zigbee_home`
