+++ cmd/zigbee_home/firmware/build_test.go
