+++ cmd/zigbee_home/firmware/flash.go
