+++ cmd/zigbee_home/firmware/flasher/mcuboot.go
