+++ cmd/zigbee_home/firmware/flasher/west.go
