+++ cmd/zigbee_home/firmware/root.go
